poc/pyproject.toml
[tool.poetry]
name = "poc"
version = "0.1.0"
description = ""
authors = ["Mikuláš Poul <mikulas.poul@xelix.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
Django = "^5.0.6"
django-pgviews-redux = "^0.10.0"
djangorestframework = "^3.15.1"
psycopg = "^3.1.19"
factory-boy = "^3.3.0"
tqdm = "^4.66.4"
django-debug-toolbar = "^4.4.2"
django-extensions = "^3.2.3"
ipython = "^8.25.0"
django-filter = "^24.2"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
